Clean up cache before insmod vrouter
According to [1] vrouter uses kmalloc. This means there should be a
300MB of contiguous memory block. Dropping cache is official
recommendation for Juniper according to [2]. That increases a chance to
have such chunk. It's very important for compute nodes with small amount
of RAM or nodes with a lot of allocated resources for other purposes
(ctl + ntw) combination for instance.

[1] https://github.com/Juniper/contrail-vrouter/blob/master/linux/vrouter_mod.c#L111-L119
[2] https://github.com/Juniper/contrail-controller/wiki/Vrouter-Module-Parameters

Change-Id: Ieff1e6cd7aa90ea2dbc66384e7f4df35a5d666ca
Signed-off-by: Sergii Golovatiuk <sgolovatiuk@mirantis.com>