IF-MAP configuration changes
OC32 implements internal IF-MAP server to replace irond.
This patch adds support for such configuration.

Monitoring changes to telegraf, collectd, prometheus and heka
to disable check of running irond process when the IF-MAP
engine is set to internal.

Enable a possibility to set IF-MAP server for contrail-control
statically instead of using contrail discovery service.

PROD-18175

Change-Id: Ic30e8e8c0c276eecba15177201fa4b86e4665305