contrail-snmp-collector auth issue in version 3
snmp collecter requires keystone configuration to be able
to connect to api with multi_tenancy enabled.

snmp collector in version 3 does not use contrail-keystone-auth.conf
by default in contradiction to what is written inside of [KEYSTONE]
section in contrail-snmp-collector.conf. Version 4.0 does read this
file and this configuration is generated for analytics nodes.

Fix this by generating configuration inside of the [KEYSTONE] section
of contrail-snmp-collector.conf configuration file.

PROD-17823

Change-Id: Ic07190efa3c9a89b6469a6a81698c0938abc9c85